requirements.txt
streamlit
mem0-ai
openai
qdrant-client